Update params passed to post url
The second error you would have encountered is that you were only
passing an object to the POST url. This will result in a url that looks
like `/posts.1` when what you really want is to pass an id param set to
the id of the post like `/posts?id=1`. It's a small change, but makes a
huge difference.